Try to log location from which the function was called
Instead of passing the full continuation mark set, convert it to
context. Extract the first relevant (meaning, skip our own calls like
`apply-traced`) item as the called-from location.

Unfortunately continuation-mark-set->context is not very precise.
Often the "call site" will be the entire enclosing function
definition, in whose body the specific function call is located.

That's unfortunate because I wanted to explore a Racket Mode feature
that would let you step through trace logs at the call sites. It could
show the calls "in situ", and feel more like a (simplified) step
debugger -- but without the "weight" of a full step debugger.

How to be more specific?

- We could do use errortrace to annotate for stack traces. This will
produce better results, but, it means whatever thing is running the
program needs to install a custom eval handler. That's not something
in our control as a library. Plus, it slows program execution for
/all/ code.

- Redefine #%app somehow?? But has some impact on all applications, if
only a conditional.

- We could change forms like `trace-define` so that they expand to
defining syntax not (just) a function. i.e. (trace-define (foo) __)
means that `foo` is syntax that records its site then calls the user's
function, and defines that function. Although this avoids impacting
the program overall, what about the consequences of `foo` being syntax
not a function? e.g. User can't use `foo` with `map` anymore? Is there
a way to keep `foo` defined as a function?