Provide srcloc for function "signature"
The intended use is for a "step tracer" sort of UI to show a call at
the definition site.

What this means varies by form:

For "(define (foo x) x)" a call "(foo 42)" should replace the
"(foo x)".

For "(let loop ([x 12]) __)" the entire span "loop ([x 12])" is the
srcloc to use.