Expand trace-define to trace-lambda ourselves
The previous commit had a hack where we stuff a property on the
trace-lambda stx given to normalize-definition, so that it flows
through into all nested lambdas.

Instead, handle expansion of trace-define to trace-lambda ourselves.
That way we can just directly supply the identifier using the #:name
option.

Also in this commit, I took time to try the style of capitalizing
syntax pattern variables.