Use chaperone-procedure to make tracing wrapper procs
An advantage is that racket/class allows method definitions to be
chaperone-procedure forms. As a result, we can supply forms like
trace-define/public. This commit only defines that, and
trace-define/override, as an initial POC.

A drawback is that it was trickier for me to figure out how to do tail
call detection and level tracking. A chaperone wrapper proc does not
call the original proc itself. So the racket/trace technique can't be
used. I /think/ I figured this out, but I'm not yet confident it works
for all cases, and don't have a rigorous understanding of how/why it
works.