Add some more examples
I was using the new examples to compare tail calls, as reported by
check-syntax, for equivalent source as handled by racket/base,
racket/trace, and vestige.

Vestige doesn't match them. :( I need to explore this more -- e.g. is
this due to the recent switch to chaperone-procedure, only? But I need
to task switch to something else, so committing just this, for now.