Refactor
Start to change this into a library that is about logging generally,
and optionally also tracing.

The default log-xxx forms created by define-logger put
current-continuation-marks in the logger event vector's data slot.
Previously, we had been /replacing/ that with our own data, and, that
data was only about tracing.

Instead: Attach any information as extra continuation marks. Plain
receivers still work (if they expect the continuation mark set in the
data slot). Special receivers can look up the extra marks.

Furthermore, recognize that some information may be useful for
non-tracing logging messages. For example, the notion of "depth". The
tracing forms set a depth. It would be nice if a plain old log-message
could use that depth to indent, too. Furthermore, a form like
`with-more-logging-depth` could increase the depth again, for any
log-xxx uses within the dynamic extent of that form. So, we get the
notion of logger "groups" or "spans", which is useful on its own, as
well as cooperating nicely with tracing depths when that is also used.

As of this commit, the docs still need to be rewritten. Probably still
want to add some conveniences for writing a log receiver.