Have with-more-logging-info record its own srcloc
Commit 34b311a attempted to overload the "caller" srcloc for this
purpose. But that's not right. For one thing, it means a logging event
not within the dyn ext of any with-more-logging-info will say that its
"caller" is some traced function, which isn't really correct. Also a
tool may just highlight that caller and nothing else (since the log
event doesn't have any tracing info).

Instead have with-more-logging-info record a distinct srcloc, along
with the things it already records like time and thread.

And have tools (e.g. Racket Mode) look for and handle that distinctly.

Note: This probably would allow us to rewrite trace-expression as a
simple composition of with-more-logging-info and log-vestige-debug.
The "show at a depth" aspect has been present for quite a few commits
now. And as of this commit, the "show the source site" aspect should
just work, too. trace-expression no longer need expand to
trace-lambda. This would probably also simplify tools; the WIP Racket
Mode implementation had to be careful handling showing caller and
called at the exact same location. Furthermore it would de-clutter the
logging, resulting in only 1 logging event not 2 per trace-expression.