Various changes
- Use one syntax property to associate values with the identifier
syntax.

- Retrieve the values, once, in make-chaperone-wrapper-procedure, as
opposed to every time a logging event occurs. Pass the values as
explicit arguments to the log-args and log-results functions. i.e.
Move some more work from run-time to compile-time.

- To the existing two values, for formals srcloc and header srcloc,
add a new one that describes positional parameters -- either a list of
symbols, or, a single symbol for a rest argument. Use this to log the
positional parameter names as well as the actual arguments. (We're
already doing this with keywords and keyword arguments, and it can be
helpful to see the names for positional parameters, too.)