Stop using syntax property
An earlier commit stopped using a syntax property to carry the
"expression datum" for trace-expression forms, because
trace-expression is gone; instead now we have log-expression.

Another commit consolidated the two remaining syntax properties into
one. (And also, retrieved the property values at wrapper creation
time, not later when creating each logger event.)

This commit removes even that one syntax property. The two special
pieces of information -- the syntax to use for the srcloc of the
"formals" and "header" -- is now just directly passed to a new, core
macro: do-trace-lambda.

(In my defense, this project has evolved from a starting point which
was racket/trace. Its trace-lambda form has an optional #:name.
Putting syntax properties on that identifier syntax was a natural way
to start. But as things have evolved, this became pointless.)

In any case, at the moment the only surface forms that even needs this
option is `trace-let`. For the other forms, it is perfectly fine to
default the "header" to the formals list, and the "formals" to its
interior elements.