Simplify trace-define
Simplify in two ways:

There is no need to build a list of formals and reverse it. Just use
recursion in the quasisyntax template.

Stop bothering to make distinct names for curried functions. Whatever
the value was, before, of naming them with their parameters, it's moot
ever since the recent change where traced calls show the parameter
names as well as the actual arguments.