continuation-mark-set->list => continuation-mark-set->iterator
For efficiency iterate until we have a number value (don't get list of
all marks beyond what we need).

Also: Change tracing logging to start with depth 0. (It had been that
way originally, then I'd switched it to 1, but I'm going back to 0. It
depends whether you want the presentation to be "a call is about to
happen" vs. "the call has happened", and how that relates to logging
that happens during the call. I think 0 is right.)