Various organizational changes
Change to multi-collection package. This seemed to be necessary for
the following item:

Change vestige/tracing/explicit to vestige/tracing.

Change vestige/tracing/class/explicit to vestige/tracing/class.

The /implicit flavors have the same names.

Make the docs multi-page.

Move the use cases "editorial" subsection to the end of the section.