Various changes
Change handling of "caller". It is only reported as part of tracing
data. Furthermore, the caller is #f unless it is the immediate caller
of the traced proc. In conjunction with that, move vestige/app from
"Logging" to "Tracing" in the docs. (Note that with-more-logging-info
records srcloc, which is used when available for non-tracing cases.)

To support that, have make-chaperone-wrapper-proc accept the wrapped
proc as a new argument.

Add a contract for make-chaperone-wrapper-proc.

Change "srcloc-as-list?" from a concept in the docs, into an actual
srcloc-as-list/c contract. Although I'm not sure how useful it is as a
contract --- as contracts go, it is very permissive -- making it an
actual binding certainly improves its treatment in the Scribble docs,
e.g. it's now a live link. And it has some utility as part of the
contract for make-chaperone-wrapper-proc.