Various changes
Add srcloc for the entire definition form. In add-presentation-sites,
use that as the location after which to show the result. After
dogfooding, I think generally this is a better
presentation. (Previously we used the header. I think this is now not
used at all, but for now I'm leaving it in as available information.)

Change the tracing 'caller mapping value from a srcloc to `(cons
immediate? srcloc)`, where immediate? is true when the caller is the
immediate caller of the traced function (which is not the case if the
immediate caller did not use tracing-#%app). Use this value in
add-presentation-sites: Only show a result for the caller when it is
immediate.

Note that the previous change means trace-case-lambda currently never
shows an immediate caller, because it introduces intermediate lambdas
for the clauses. Need to fix that.