Stop using chaperone-procedure
Go back to using a direct wrapper procedure.

The original motivation for switching to chaperone-procedure was
racket/class. At first that seemed like the only way (wrong). Plus I
thought maybe chaperone-procedure is the idiomatic way to install
"advice" in Racket. And finally, perhaps the racket/class special
handling of chaperone-procedure, which AFAICT was motivated by
racket/contract, would be something I'd discover elsewhere; maybe
racket/contract would have "paved the way".

However I got tired of struggling with complications like depth
tracking and tail detection, including needing a dubious work-around
for <https://github.com/racket/racket/issues/1836>.

I came up with an initial, stupid, slow way to make this work (for
some definition of "work") for racket/class. I plan to loop back later
to make it work better, as well as test with more actual racket/class
example code. Anyway, my initial thought was that racket/class was
just a nice-to-have, not a must-have.

Also: Stop passing a `depth` argument to log-args and log-results.
This was always a point of confusion, because it was only used to
format the legacy logger message (e.g. for number of > or < chars).
For everything else, depth was determined by the continuation mark
value. Now log-args and log-results look that up for the legacy
message. (This is also faster now that it's a simple
continuation-mark-set-first, as opposed to needing to look at some
sequence like '(app-mark app-mark 1 2) as was the case with the
chaperone approach.)