More context changes
Work around Racket struct generativity. Elsewhere we already we avoid
srcloc structs when we record information. In this case, we're getting
things "pre-baked" from continuation-mark-set->context, so we need
to use struct->vector and so on.

Adjust the path exclusion to be just our private dir, so that the
context is shown for example.rkt.