Move more work from logging time to receiving time
Specifically, record data in vectors when logging, and wait until
receiving to produce hash-tables.

Although I'm not sure of the extent of impact, in principle this
should save some time and allocation. As much as possible and
practical, we want to defer the impact of logging and tracing
to receiving time.